Mass update: use long instead of int on every table id

+++ DataAccess/Data/EmployeeManagement/Contracts/IEmployeeGovtIdCardData.cs
@@ -12,7 +12,7 @@ namespace DataAccess.Data.EmployeeManagement.Contracts
     {
         List<EmployeeGovtIdCardModel> GetAllByEmployeeNumber(string employeeNumber);
 
-        List<EmployeeGovtIdCardModel> GetAllByGovtAgency(int govtAgencyId);
+        List<EmployeeGovtIdCardModel> GetAllByGovtAgency(long govtAgencyId);
 
         EmployeeGovtIdCardModel GetByEmployeeIdNumber(string employeeIdNumber);
     }
