+++ DataAccess/Data/InventoryManagement/Contracts/IIngredientInventoryData.cs
@@ -10,9 +10,9 @@ namespace DataAccess.Data.InventoryManagement.Contracts
 {
     public interface IIngredientInventoryData : IDataManagerCRUD<IngredientInventoryModel>
     {
-        decimal GetRemainingQtyValueByIngredient(int ingredientId);
-        List<IngredientInventoryModel> GetAllByIngredient(int ingredientId);
-        IngredientInventoryModel GetByIdAndIngredient(int ingredientId, long id);
+        decimal GetRemainingQtyValueByIngredient(long ingredientId);
+        List<IngredientInventoryModel> GetAllByIngredient(long ingredientId);
+        IngredientInventoryModel GetByIdAndIngredient(long ingredientId, long id);
         List<IngredientInventoryModel> GetAllByExpirationDateRange(DateTime startDate, DateTime endDate);
     }
 }
