+++ EntitiesShared/InventoryManagement/Models/ProductIngredientInventoryDeduction.cs
@@ -8,7 +8,7 @@ namespace EntitiesShared.InventoryManagement.Models
 {
     public class ProductIngredientInventoryDeduction
     {
-        public int IngredientId { get; set; }
+        public long IngredientId { get; set; }
 
         public long IngredientInventoryid { get; set; }
